Some `hexagon-linux` fixes + enable std tests in CI

+++ lib/std/os/linux/hexagon.zig
@@ -170,7 +170,7 @@ pub const Flock = extern struct {
 
 pub const blksize_t = i32;
 pub const nlink_t = u32;
-pub const time_t = i32;
+pub const time_t = i64;
 pub const mode_t = u32;
 pub const off_t = i64;
 pub const ino_t = u64;
